signin part 1
there is now an endpoint to sign someone in
the table does not yet exist
and there are more holes in functionality
i also added routes though

i need to add a lot of localStorage shit

and actual user TOOLS. like, why sign in lol

oh and production is still broken because idfk why. directory failures?